requirements.txt
numpy
pandas
sklearn.manifold